% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.R
\name{gg_circle}
\alias{gg_circle}
\title{Function to add circles and ellipses to ggplot images.}
\usage{
gg_circle(rx, ry, xc, yc, color = "black", fill = NA, ...)
}
\arguments{
\item{rx}{parameter.}

\item{ry}{parameter.}

\item{xc}{parameter.}

\item{yc}{parameter.}

\item{color}{parameter.}

\item{fill}{parameter.}
}
\description{
`gg_circle`` produces circles and ellipses that can be added to ggplot images.
}
