% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.R
\name{ropls.enrich}
\alias{ropls.enrich}
\title{Perform FGSEA enrichment.}
\usage{
ropls.enrich(var, var.name, value.name, filterset = FALSE)
}
\arguments{
\item{var}{data frame containing a variable column, at least one variable information column, and a score column (e.g. p-value, fold-change, OPLS-DA loading).}

\item{var.name}{The name of the variable column in quotes, must match the variable column name exactly (e.g. "Precursor.Ion.Name").}

\item{value.name}{The name of the value column in quotes, must match the value column name exactly (e.g. "p-value", "fc" or "p1").}

\item{filterset}{If FALSE (default) all sets are used for enrichment, or could be used to use a particular set, matching the name of a variable information column (e.g. "class", "cl", "uns")}
}
\description{
`ropls.enrich`` performs FGSEA enrichment on a dataframe containing variables, variable information and variable scores.
The input of the function is a dataframe with a variable column, one or multiple columns containing variable information and scores.
The variable column may contain non-unique names, but will be filtered for distinct entries.
The variable information may be one or multiple columns, and can be selected using the filterset variable.
}
