% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.R
\name{ropls.plot}
\alias{ropls.plot}
\title{Produce publication-quality ggplot images of ropls objects.}
\usage{
ropls.plot(d, plottype = "score", xvar, yvar, hotelling = FALSE,
  ellipse = FALSE, col.var = NULL)
}
\arguments{
\item{d}{ropls object created by [opls()].}

\item{plottype}{Which type of plot to produce. Options include scores (default), loadings and metrics.}

\item{xvar}{which component to plot on the x-axis. Must refer to a single component ("p1") or orthogonal ("o1") component present in the supplied ropls object.}

\item{yvar}{which component to plot on the y-axis. Must refer to a single component ("p1") or orthogonal ("o1") component present in the supplied ropls object.}

\item{hotelling}{If set to TRUE, will add Hotelling's T2 ellipse to the plot.}

\item{ellipse}{If set to TRUE, will add an ellipse around the groups supplied in the model. Does not work for PCA, PLS and OPLS models.}

\item{col.var}{Vector of categorical or numerical values used to colour the variables of the loading plot.}
}
\description{
`ropls.plot`` produces publication-quality ggplot images of ropls objects.
The functions automatically determines the model type ("PCA", "PLS", "OPLS","PLS-DA","OPLS-DA") from the ropls object.
The components shown in the plot are flexible and determined by the xvar and yvar parameters.
}
