% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.R
\name{gg_circle}
\alias{gg_circle}
\title{Function to add circles and ellipses to ggplot images.}
\usage{
gg_circle(rx, ry, xc, yc, color = "black", fill = NA, ...)
}
\arguments{
\item{rx}{paramater.}

\item{ry}{paramater.}

\item{xc}{paramater.}

\item{yc}{paramater.}

\item{color}{paramater.}

\item{fill}{paramater.}
}
\description{
gg_circle produces circles and ellipses that can be added to ggplot images.
}
